fix: create wallet with descriptor=false
